Permisos aplicados en todos los modulos

+++ src/Pages/Ventas/ListaGarantia.jsx
@@ -28,8 +28,16 @@ import axios from 'axios';
 //GENERADOR DE PDF
 import { generatePDF } from '../../Components/generatePDF';
 
-export const ListaGarantia = ({props,data,update}) => {
-
+export const ListaGarantia = ({idRol,data,update}) => {
+  const [permisos, setPermisos] = useState([]);
+  const urlPermisos = 'http://localhost:3000/api/permiso/consulta'
+  const dataPermiso={
+    idRol:idRol,
+    idObj:8
+  }
+  useEffect(()=>{
+    axios.post(urlPermisos,dataPermiso).then((response)=>setPermisos(response.data))
+  },[])
   const [marcah, setMarcah] = useState()
   const [cambio, setCambio] = useState(0)
 
@@ -49,28 +57,33 @@ export const ListaGarantia = ({props,data,update}) => {
 
   //IMPRIMIR PDF
 const handleGenerarReporte = () => {
-  const formatDataForPDF = () => {
-    const formattedData = tableData.map((row) => {
-      const fechaCre = new Date(row.fechaNacimiento);
-      const fechaNacimiento = String(fechaCre.getDate()).padStart(2,'0')+"/"+
-                            String(fechaCre.getMonth()).padStart(2,'0')+"/"+
-                            fechaCre.getFullYear();
-                            return {
-                              'N°':row.IdGarantia,
-                              'Descripción':row.descripcion, 
-                              'Meses de Garantia':row.Meses, 
-                              'Producto':row.producto, 
-                              'Estado':row.estado,                        
-                            };
-    });
-    return formattedData;
-  };
+  if (permisos[0].consultar==="n") {
+    swal("No cuenta con los permisos para realizar esta accion","","error")
+  } else {
+    const formatDataForPDF = () => {
+      const formattedData = tableData.map((row) => {
+        const fechaCre = new Date(row.fechaNacimiento);
+        const fechaNacimiento = String(fechaCre.getDate()).padStart(2,'0')+"/"+
+                              String(fechaCre.getMonth()).padStart(2,'0')+"/"+
+                              fechaCre.getFullYear();
+                              return {
+                                'N°':row.IdGarantia,
+                                'Descripción':row.descripcion, 
+                                'Meses de Garantia':row.Meses, 
+                                'Producto':row.producto, 
+                                'Estado':row.estado,                        
+                              };
+      });
+      return formattedData;
+    };
+  
+    const urlPDF = 'Report_Garantias.pdf';
+    const subTitulo = "LISTA DE GARANTIAS"
+  
+    const orientation = "landscape";
+    generatePDF(formatDataForPDF, urlPDF, subTitulo, orientation);
+  }
 
-  const urlPDF = 'Report_Garantias.pdf';
-  const subTitulo = "LISTA DE GARANTIAS"
-
-  const orientation = "landscape";
-  generatePDF(formatDataForPDF, urlPDF, subTitulo, orientation);
 };
 
   const navegate = useNavigate();
@@ -113,41 +126,46 @@ const handleGenerarReporte = () => {
 
   //FUNCION DE ELIMINAR 
   function handleDel(id) {
-    swal({
-      content: (
-        <div>
-          <div className="logoModal">¿Desea Elimiar esta garantia?</div>
-          <div className="contEditModal">
+    if (permisos[0].eliminar === "n") {
+      swal("No cuenta con los permisos para realizar esta accion","","error")
+    } else {
+      swal({
+        content: (
+          <div>
+            <div className="logoModal">¿Desea Elimiar esta garantia?</div>
+            <div className="contEditModal">
+            </div>
           </div>
-        </div>
-      ),
-      buttons: {
-        cancel: 'Eliminar',
-        delete: 'Cancelar',
-      },
-    }).then(async(op) => {
-
-      switch (op) {
-        case null:
-
-          let data = {
-            IdGarantia: id,
-          };
-    
-          console.log(data);
-    
-          await axios.delete(urlDelGarantias,{data}).then(response=>{
-            swal("Garantia eliminada correctamente","","success")
-            setCambio(cambio+1)
-          }).catch(error=>{
-            console.log(error);
-            swal("Error al eliminar la garantia","","error")
-          })
-        break;
-        default:
-        break;
-      }
-    });
+        ),
+        buttons: {
+          cancel: 'Eliminar',
+          delete: 'Cancelar',
+        },
+      }).then(async(op) => {
+  
+        switch (op) {
+          case null:
+  
+            let data = {
+              IdGarantia: id,
+            };
+      
+            console.log(data);
+      
+            await axios.delete(urlDelGarantias,{data}).then(response=>{
+              swal("Garantia eliminada correctamente","","success")
+              setCambio(cambio+1)
+            }).catch(error=>{
+              console.log(error);
+              swal("Error al eliminar la garantia","","error")
+            })
+          break;
+          default:
+          break;
+        }
+      });
+    }
+   
   };
 
   //BOTON DE RETROCEDER 
@@ -157,27 +175,32 @@ const handleGenerarReporte = () => {
 
    //FUNCION DE ACTUALIZAR 
   function handleUpdt(id) {
-    swal({
-      buttons: {
-        update: 'Actualizar',
-        cancel: 'Cancelar',
-      },
-      content: (
-        <div className="logoModal">
-          ¿Desea actualizar la Garantia: {id.descripcion}?
-        </div>
-      ),
-    }).then((op) => {
-      switch (op) {
-          case 'update':
-          data(id)
-          update(true)
-      navegate('/menuVentas/RegistroGarantia')
-      break;
-      default:
-      break;
-      }
-    });
+    if (permisos[0].actualizar === "n") {
+      swal("No cuenta con los permisos para realizar esta accion","","error")
+    } else {
+      swal({
+        buttons: {
+          update: 'Actualizar',
+          cancel: 'Cancelar',
+        },
+        content: (
+          <div className="logoModal">
+            ¿Desea actualizar la Garantia: {id.descripcion}?
+          </div>
+        ),
+      }).then((op) => {
+        switch (op) {
+            case 'update':
+            data(id)
+            update(true)
+        navegate('/menuVentas/RegistroGarantia')
+        break;
+        default:
+        break;
+        }
+      });
+    }
+   
   };
 
   return (
@@ -212,7 +235,12 @@ const handleGenerarReporte = () => {
             <Button
               className="btnCreate"
               onClick={() => {
-                navegate('/menuVentas/RegistroGarantia');
+                if (permisos[0].insertar === "n") {
+                  swal("No cuenta con los permisos para realizar esta accion","","error")
+                } else {
+                  navegate('/menuVentas/RegistroGarantia');
+                }
+                
               }}
             >
               <AddIcon style={{ marginRight: '5px' }} />
